Consolidate sorters and add additional sorters
